+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.55 (Pending):
+* Remove PersistenceTestModule.registerUserTransaction() as it is unused.
 * Remove support for deriving the tables to clean in PersistenceTestModule from @Table annotations as
   domgen now explicitly lists all the tables to clean.
 * Always inject EntityListeners in PersistenceTestModule as no longer need to support EE6.